va-claims-ai-review/backend/requirements.txt
# VA Claims AI Document Review System - Backend Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
pypdfium2==4.25.0
pdfplumber==0.10.3

# OCR and Image Processing
opencv-python==4.8.1.78
easyocr==1.7.1

# NLP and AI
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# LLM Integration
anthropic==0.8.0
openai==1.6.1
langchain==0.0.350
tiktoken==0.5.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
pgvector==0.2.4
redis==5.0.1

# API and Integration
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4

# Monitoring and Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
dateparser==1.2.0
rapidfuzz==3.5.2